Revert some of the structural changes of the project
