+++ f4pga/common_modules/analysis.py
@@ -19,8 +19,8 @@
 
 from pathlib import Path
 
-from f4pga.flows.common import vpr_specific_values, vpr as common_vpr, VprArgs
-from f4pga.flows.module import Module, ModuleContext
+from f4pga.common import vpr_specific_values, vpr as common_vpr, VprArgs
+from f4pga.module import Module, ModuleContext
 
 
 def analysis_merged_post_implementation_file(ctx: ModuleContext):
