+++ f4pga/common_modules/fasm.py
@@ -19,8 +19,8 @@
 
 from pathlib import Path
 
-from f4pga.flows.common import vpr_specific_values, VprArgs, get_verbosity_level, sub as common_sub
-from f4pga.flows.module import Module, ModuleContext
+from f4pga.common import vpr_specific_values, VprArgs, get_verbosity_level, sub as common_sub
+from f4pga.module import Module, ModuleContext
 
 
 class FasmModule(Module):
