+++ f4pga/common_modules/place_constraints.py
@@ -19,8 +19,8 @@
 
 from pathlib import Path
 
-from f4pga.flows.common import sub as common_sub
-from f4pga.flows.module import Module, ModuleContext
+from f4pga.common import sub as common_sub
+from f4pga.module import Module, ModuleContext
 
 
 class PlaceConstraintsModule(Module):
