+++ f4pga/common_modules/route.py
@@ -19,8 +19,8 @@
 
 from pathlib import Path
 
-from f4pga.flows.common import vpr_specific_values, vpr as common_vpr, VprArgs, options_dict_to_list, save_vpr_log
-from f4pga.flows.module import Module, ModuleContext
+from f4pga.common import vpr_specific_values, vpr as common_vpr, VprArgs, options_dict_to_list, save_vpr_log
+from f4pga.module import Module, ModuleContext
 
 
 def route_place_file(ctx: ModuleContext):
