--- f4pga/flows/requirements.txt
@@ -1,2 +1,0 @@
-colorama
-pyyaml
