+++ f4pga/module_inspector.py
@@ -19,8 +19,8 @@
 
 from colorama import Style
 
-from f4pga.flows.module import Module
-from f4pga.flows.common import decompose_depname
+from f4pga.module import Module
+from f4pga.common import decompose_depname
 
 
 def _get_if_qualifier(deplist: "list[str]", qualifier: str):
