+++ f4pga/requirements.txt
@@ -1,1 +1,2 @@
--r ./flows/requirements.txt
+colorama
+pyyaml
