+++ f4pga/setup.py
@@ -78,12 +78,10 @@ setuptools_setup(
     author="F4PGA Authors",
     description="F4PGA.",
     url="https://github.com/chipsalliance/f4pga",
-    packages=["f4pga", "f4pga.flows", "f4pga.flows.common_modules", "f4pga.wrappers.sh", "f4pga.wrappers.tcl"],
+    packages=["f4pga", "f4pga.common_modules", "f4pga.wrappers.sh", "f4pga.wrappers.tcl"],
     package_dir={"f4pga": "."},
     package_data={
-        "f4pga.flows": [
-            "*.yml",
-        ],
+        "f4pga": ["part_db.yml", "platforms.yml"],
         "f4pga.wrappers.sh": ["xc7/*.f4pga.sh", "quicklogic/*.f4pga.sh"],
         "f4pga.wrappers.tcl": [
             "xc7/*.f4pga.tcl",
